refactor(auth): prefer `super` to `crate`
In accordance to Enarx style guide https://github.com/enarx/enarx/blob/main/STYLE.md#import-paths
