+++ crates/auth/src/builder.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Profian Inc. <opensource@profian.com>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::providers;
+use super::providers;
 
 use axum::routing::get;
 use axum::{Extension, Router};
