+++ crates/auth/src/redirect.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Profian Inc. <opensource@profian.com>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::providers::github;
+use super::providers::github;
 
 use axum::response::{self, IntoResponse, Response};
